Moved hero states to separate classes
